+++ types/entity.d.ts
